backend/requirements.txt
# Core Flask web framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Deep Learning and AI Models
deepface>=0.0.79
mediapipe>=0.10.0

# TensorFlow (required by deepface) - using more recent version for Python 3.13
tensorflow>=2.16.0

# Image processing and utilities
scikit-image>=0.22.0
matplotlib>=3.8.0

# Data handling and utilities
pandas>=2.1.0
requests>=2.31.0

# Logging and configuration
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-flask>=1.3.0

# Additional computer vision libraries (optional, may not work on all systems)
# dlib>=19.24.0
# face-recognition>=1.3.0

# Color analysis
# colorspacious>=1.1.0
webcolors>=1.13

# Performance optimization
# numba>=0.58.0 